moved about sections into components
