+++ src/components/about/AboutMe.tsx
@@ -1,0 +1,18 @@
+import AboutContent from './AboutContent'
+
+export default function AboutMe() {
+  return (
+    <AboutContent header='About Me'>
+      I didn't take the typical path to becoming a software developer. As a kid I was always curious about how things
+      were built and was good at math and science. I even 'dissected' an old computer in my bedroom, but in reality it
+      was more of a demolition. After high school, I went to an engineering focused college in hopes of being a Civil
+      Engineer, but quickly found it didn't interest me. I changed my focus to business because I've always liked math
+      and my new department required me to take an Information Systems class. There I found my passion for computers and
+      how they work. My first programming language was Java and I loved tinkering around and making little terminal
+      applications. My Java skills grew further at my first two jobs, but my love for stopped. In my own time I tought
+      myself JavaScript, HTML, and CSS building small sites and single-purpose applications. From there I learned React
+      and TypeScript and that helped me land to my current job at Coolfire Solutions. Today, I enjoy the work that I do
+      for Coolfire and am still learning new things outside of work.
+    </AboutContent>
+  )
+}
