+++ src/components/about/index.ts
@@ -1,3 +1,6 @@
 export { default as AboutContent } from './AboutContent'
 export { default as AboutList } from './AboutList'
+export { default as AboutMe } from './AboutMe'
 export { default as ContactMe } from './ContactMe'
+export { default as PastProjects } from './PastProjects'
+export { default as WorkExperience } from './WorkExperience'
